[level 2] Title: n^2 배열 자르기, Time: 8.46 ms, Memory: 113 MB -BaekjoonHub

+++ 프로그래머스/lv2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -25,7 +25,7 @@ class Solution {
         int idx = 0;    // answer 배열 인덱스
         // for문
         for(long i = left; i <= right; i++){
-            answer[idx++] = (long)(Math.max((long)(i / n), (long)(i % n)) + 1);
+            answer[idx++] = Math.max((i / n), (i % n)) + 1;
         }
         
         return answer;
